DEM: Remove accidental FP64 math from the sphere contact force kernels
The sphere-sphere normal force helpers computed the contact distance in
double (int3_to_double3, a double division and a double rsqrt) and then
rounded the result to float, and many single-precision expressions in the
contact, boundary and integrator kernels used unsuffixed double literals
(2., 0.5, 2.0 / 3.0, sqrt(5.0 / 6.0), log(EPSILON)) or double unit
fields. On consumer GPUs FP64 runs at a small fraction of the FP32 rate,
so this made computeSphereContactForces and the frictionless kernels
largely FP64 bound.

Compute these in float with float literals and intrinsics. Precision is
unchanged in practice: positions are integers in SU, the double path
already rounded reciplength to float, and the float rounding error of the
penetration is far below one SU. The double-precision distance in the
unused material-based sphere BC path is left as is.

+++ src/chrono_dem/gpu/ChDemHelpers.cuh
@@ -224,7 +224,7 @@ inline __device__ float3 computeRollingAngAcc(ChSystemDem_impl::GranSphereDataPt
                 // convert v_rot to user unit for threshold comparison
                 // float v_rot_uu = v_rot_su * velo_su2uu;
 
-                if (v_rot_su < 1e-4f * gran_params->TIME_UNIT / gran_params->LENGTH_UNIT) {
+                if (v_rot_su < 1e-4f * (float)gran_params->TIME_UNIT / (float)gran_params->LENGTH_UNIT) {
                     return make_float3(0.f, 0.f, 0.f);
                 }
 
@@ -358,8 +358,8 @@ inline __device__ float3 computeFrictionForces_matBased(ChSystemDem_impl::GranPa
 
     computeMultiStepDisplacement(gran_params, sphere_data, contact_index, vrel_t, contact_normal, delta_t);
     // evaluate kt and gt
-    float kt = 8. * G_eff * sqrt_Rd;
-    float gt = -2. * beta * std::sqrt(5. / 6. * m_eff * kt);
+    float kt = 8.f * G_eff * sqrt_Rd;
+    float gt = -2.f * beta * sqrtf(5.f / 6.f * m_eff * kt);
 
     // float3 tangent_force =  - kt * delta_t - gt * vrel_t; // this works for sph-sph and sph-bc contact
